requirements.txt
pandas
redis
numpy
bcrypt==3.2.0
celery==5.2.3
email-validator==1.1.3
Flask
Flask-Caching==1.10.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==4.1.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
SQLAlchemy==1.4.32
werkzeug==2.0.0
requests
python-dateutil
gunicorn==20.0.4